refactor(filter-panel-utilities): move legacy utilities to directory
