+++ src/components/FilterPanel/FilterPanel.stories.js
@@ -80,7 +80,7 @@ storiesOf(patterns('FilterPanel'), module)
         text: `## Basic implementation of Filter Panel
 
           The most crucial requirement for this component is the structure of the filter data. Be sure to
-          check out type definitions for [filter data](https://github.com/carbon-design-system/ibm-security/blob/master/src/components/FilterPanel/FilterPanelUtilities.js#L60).
+          check out type definitions for [filter data](https://github.com/carbon-design-system/ibm-security/blob/master/src/components/FilterPanel/LEGACY_FilterPanel/FilterPanelUtilities.js#L60).
           Using the unique prop for each label will override the version passed in the \`labels\` prop.`,
       },
     }
