+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterPanelUtilities.js
