+++ src/components/FilterPanel/LEGACY_FilterPanel/FilterSelector/FilterSelector.js
@@ -9,7 +9,7 @@ import React, { Fragment } from 'react';
 import { getComponentNamespace } from '../../../../globals/namespace';
 
 import Checkbox from '../../../Checkbox';
-import { filterFiltersPropTypes } from '../../FilterPanelUtilities';
+import { filterFiltersPropTypes } from '../FilterPanelUtilities';
 
 export const namespace = getComponentNamespace('filter-selector');
